--- examples/shareable.html
@@ -1,93 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <title>MapKnitter Lite</title>
-  <meta charset="utf-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"/>
-  <meta http-equiv="x-ua-compatible" content="ie=edge"/>
- 
-  <link rel="stylesheet" href="../node_modules/leaflet/dist/leaflet.css">
-  <link rel="stylesheet" href="../dist/vendor.css">
-  <link rel="stylesheet" href="../dist/leaflet.distortableimage.css" media="screen" title="no title">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-  <link rel="stylesheet" href="https://unpkg.com/leaflet-control-geocoder/dist/Control.Geocoder.css" />
-
-  <script src="../node_modules/leaflet/dist/leaflet.js" type="text/javascript" charset="utf-8"></script>
-  <script src="../dist/vendor.js"></script>
-  <script src="../dist/leaflet.distortableimage.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-  <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-  <script src="https://maps.googleapis.com/maps/api/js?key=<key>&libraries=places"></script>
-  <script src="https://unpkg.com/leaflet-control-geocoder/dist/Control.Geocoder.js"></script>
-  <script src="../dist/leaflet.distortableimage.js"></script>
-  <script src="./js/shareable.js" defer type="module"></script>
-  <script src="./js//modules/paginator.js" type="module" defer></script>
-</head>
-<body style="margin:0;">
-      
-  <div class="modal fade" id="welcomeModal" aria-labelledby="exampleModalLabel" tabindex="-1" aria-hidden="true">
-        <div class="modal-dialog modal-dialog-centered modal-dialog-scrollable">
-          <div class="modal-content bg-light">
-            <div class="modal-header d-block">
-              <div class="d-flex justify-content-between align-items-center">
-                <h2 class="modal-title">Welcome to MapKnitter Lite</h2>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close" ></button>
-              </div>
-              <p> <small>(while MapKnitter is offline)</small> </p>
-
-            </div>
-            <div class="modal-body">
-              <p>First, you’ll need to go to <a href="https://archive.org/" target="_blank">archive.org</a>, create an account, and then upload your images.
-                <a href="https://archive.org/embed/help-upload-overview" target="_blank">Watch this helpful video guide.</a>
-              </p>
-              <p>You’ll get an address that looks like this:
-              <a href="https://archive.org/details/texas-barnraising/" target="_blank">
-                https://archive.org/details/texas-barnraising/
-              </a>
-            </p>
-            <p>Paste it here to begin:</p>
-            <form id="form">
-              <input id="input" type="text" class="form-control" placeholder="https://archive.org/details/..." required>
-              <button type="submit" class="btn btn-primary mt-4 mb-5">Begin</button>
-            </form>
-            <p>
-                <a href="local.html">Click here</a> to try this offline with local files only.
-              </p>
-          </div>
-        </div>
-      </div>
-  </div>
-    
-  <i title="Open Sidebar" id="mapToggle" class="fa fa-bars fa-3x " style="position: absolute; right: 0; top:30px; margin: 1rem; z-index: 900; color: white; cursor: pointer;" aria-hidden="true"></i>
-    
-  <div id="map" style="width:100%; height:100%; position:absolute; top:0;"></div>
-
-  <div class="offcanvas offcanvas-end" data-bs-backdrop="false" data-bs-keyboard="false" tabindex="1" id="offcanvasRight" aria-labelledby="offcanvasRightLabel">
-      <div class="offcanvas-header">
-        <h3 id="offcanvasRightLabel">Images</h3>
-        <button type="button" id="restoreWelcomeModalBtn" class="btn btn-primary btn-sm" data-bs-dismiss="offcanvas" aria-label="Close">Get Image</button>
-        <button type="button" class="btn btn-primary btn-sm" data-bs-dismiss="offcanvas" aria-label="Close">View Map</button>
-      </div>
-      <hr>
-      <div class="offcanvas-body">
-        <p id="response" class="mb-5">No link = No images...</p>
-        <div id="imgContainer"></div>
-      </div>
-        <hr>
-      <div class="offcanvas-footer d-flex justify-content-evenly align-items-center">
-       <button title="Previous" id="prevBtn" style="all: unset; cursor: pointer;">
-        <i class=" fa fa-regular fa-chevron-left"></i>
-       </button>
-        <p style="margin: 5px;" id="range">1-50 of 200</p>
-       <button title="Next" id="nextBtn" style="all: unset; cursor: pointer;">
-        <i  class=" fa fa-regular fa-chevron-right"></i>
-       </button>
-      </div>
-  </div>
-
-</body>
-</html>
-
-        
-
